iSEE extension for panels related to pathway analysis

iSEEpathways is a package that provides panels for iSEE, facilitating the interactive visualisation of pathway analysis results.

See also

Author

Kevin Rue-Albrecht kevin.rue-albrecht@imm.ox.ac.uk

Examples

library("iSEEde")