The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Slot overview

This class inherits all its slots directly from its parent class iSEEpathwaysResults.

Constructor

iSEEfgseaResults(data, pathwayType, pathwaysList = NULL, featuresStats = NULL) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

pathwayType

A character scalar specifying the type of pathway (e.g., "GO"). See embedPathwaysResults.

pathwaysList

A named list of pathways and associated feature identifiers.

featuresStats

Feature-level statistics used in the pathway analysis.

Supported methods

In the following code snippets, x is an instance of a iSEEfgseaResults class. Refer to the documentation for each method for more details on the remaining arguments.

embedPathwaysResults(x, se, name, pathwayType, ...) embeds x in the column name of metadata(se)[["iSEEpathways"]].

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval      padj   log2err         ES       NES size
#> 1: GO:0032467 8.655085e-05 0.4516223 0.5384341 -0.5209010 -2.031124   46
#> 2: GO:0006874 4.185815e-04 0.6263911 0.4984931  0.2865056  1.505873  303
#> 3: GO:0002758 5.914817e-04 0.6263911 0.4772708  0.3164823  1.601669  209
#> 4: GO:0055074 7.438285e-04 0.6263911 0.4772708  0.2800043  1.478614  327
#> 5: GO:0060964 7.454386e-04 0.6263911 0.4772708  0.5270390  1.935447   34
#> 6: GO:1901386 8.224892e-04 0.6263911 0.4772708 -0.6021877 -1.978540   22
#>                                  leadingEdge
#> 1:  SPAST,TAS1R2,CENPV,RHOA,CDC14B,CHMP3,...
#> 2: PLCH2,ATP2B4,CHERP,CASQ2,FKBP1A,TRPC3,...
#> 3:     PUM2,TYROBP,TYRO3,LGR4,SRC,HSPA1B,...
#> 4: PLCH2,ATP2B4,CHERP,CASQ2,FKBP1A,TRPC3,...
#> 5:   PUM2,DND1,ZC3H10,POU5F1,PUM1,ELAVL1,...
#> 6:  MIR208A,CBARP,CRHR1,OPRL1,PPP3CC,GEM,...

##
# Simulate a SummarizedExperiment object
##

se <- SummarizedExperiment()

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea",
  pathwayType = "GO", pathwaysList = pathways, featuresStats = gene_stats)
se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Access ----
##

# TODO: create getter function
metadata(se)[["iSEEpathways"]][["fgsea"]]
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.297414  0.895741 0.1142665  0.310539  1.113417        30
#> GO:0000018  0.420601  0.946474 0.0925529  0.214258  1.014846       134
#> GO:0000027  0.267161  0.887883 0.1114627 -0.335166 -1.141478        25
#> GO:0000028  0.761079  0.996120 0.0565300 -0.258580 -0.811337        19
#> GO:0000038  0.493590  0.969277 0.0833634  0.263242  0.976530        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257 0.7775424  0.996120 0.0599892  0.183261  0.888384       154
#> GO:2001258 0.2490909  0.882086 0.1147507 -0.272550 -1.125086        59
#> GO:2001259 0.5911602  0.991846 0.0664364 -0.225623 -0.941642        66
#> GO:2001267 0.0842912  0.768936 0.2139279 -0.461393 -1.390403        17
#> GO:2001279 0.9368836  0.996120 0.0480220 -0.230175 -0.664146        15
#>                        leadingEdge
#>                             <list>
#> GO:0000002   DNAJA3,SESN2,DNA2,...
#> GO:0000018    KPNA2,EPC1,EP400,...
#> GO:0000027     TRAF7,MRM2,RPL5,...
#> GO:0000028    RRP7A,ABT1,RPS27,...
#> GO:0000038   ELOVL4,PEX5,HACD4,...
#> ...                            ...
#> GO:2001257 CASQ2,FKBP1A,KCNAB1,...
#> GO:2001258 GSTO1,MIR208A,SUMO1,...
#> GO:2001259    GSTO1,NPPA,KCNE2,...
#> GO:2001267     BAX,LGALS9,JAK2,...
#> GO:2001279   AVP,MIR766,MIR132,...