The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEfgseaResults(data, row.names = rownames(data)) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

Supported methods

embedPathwaysResults(x, se, name, pathwayType, class = "fgsea", ...) embeds x in the column name of metadata(se)[["iSEEpathways"]].

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
#> 
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval      padj   log2err         ES       NES size
#> 1: GO:0006968 0.0001416741 0.7392555 0.5188481 -0.4819946 -1.962434   52
#> 2: GO:0071347 0.0004037384 0.8360929 0.4984931  0.3864373  1.779327  111
#> 3: GO:0070555 0.0005667329 0.8360929 0.4772708  0.3472104  1.660057  141
#> 4: GO:0048245 0.0007165381 0.8360929 0.4772708  0.5558281  1.971734   29
#> 5: GO:0006195 0.0008110401 0.8360929 0.4772708  0.4558411  1.820537   52
#> 6: GO:0007009 0.0010259983 0.8360929 0.4550599  0.3318156  1.620597  170
#>                                   leadingEdge
#> 1:          RELA,CCR2,CCR6,CD5L,NCF2,LSP1,...
#> 2:       EGR1,CCL18,IL1R1,HIF1A,HAS2,ZBP1,...
#> 3: YTHDC2,EGR1,CCL18,MIR101-1,IL1R1,HIF1A,...
#> 4:     CCL18,CCL3L3,CCL3,CCL15,CCL14,XCL2,...
#> 5:      NUDT3,DNPH1,PDE5A,HINT1,PNP,PDE9A,...
#> 6:   TMEFF2,NHERF1,BIN3,PRX,XKR9,BAIAP2L2,...

##
# Simulate a SummarizedExperiment object
##

se <- SummarizedExperiment()

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea", pathwayType = "GO")
se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Access ----
##

# TODO: create getter function
metadata(se)[["iSEEpathways"]][["fgsea"]]
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.265136  0.921723 0.1198878  0.318838  1.140417        30
#> GO:0000018  0.548255  0.979960 0.0756946  0.205444  0.971034       134
#> GO:0000027  0.424528  0.970057 0.0847985 -0.301334 -1.038796        25
#> GO:0000028  0.920408  0.997589 0.0503964  0.217209  0.683728        19
#> GO:0000038  0.740206  0.997034 0.0610364  0.226106  0.838822        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.850895  0.997034 0.0528050  0.175642  0.851055       154
#> GO:2001258  0.496842  0.971939 0.0822055  0.238779  0.990188        59
#> GO:2001259  0.849687  0.997034 0.0551352  0.191114  0.801009        66
#> GO:2001267  0.363636  0.953974 0.0975449  0.358048  1.081815        17
#> GO:2001279  0.240481  0.919424 0.1238422 -0.405791 -1.180479        15
#>                        leadingEdge
#>                             <list>
#> GO:0000002     PIF1,DNA2,LONP1,...
#> GO:0000018   ACTB,ZSCAN4,SHLD1,...
#> GO:0000027 RRS1,FASTKD2,RPL10L,...
#> GO:0000028        RPS19,RPS15,PWP2
#> GO:0000038 ABCD4,HACD1,SLC27A4,...
#> ...                            ...
#> GO:2001257 KCNE2,CLIC2,MIR208B,...
#> GO:2001258 KCNE2,CLIC2,MIR208B,...
#> GO:2001259  KCNE2,ABCC8,ANTKMT,...
#> GO:2001267     PARK7,IL9,CAAP1,...
#> GO:2001279   PLA2G3,CD74,PTGS2,...