The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Slot overview

This class inherits all its slots directly from its parent class iSEEpathwaysResults.

Constructor

iSEEfgseaResults(data, pathwayType, pathwaysList = NULL, featuresStats = NULL) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

pathwayType

A character scalar specifying the type of pathway (e.g., "GO"). See embedPathwaysResults.

pathwaysList

A named list of pathways and associated feature identifiers.

featuresStats

Feature-level statistics used in the pathway analysis.

Supported methods

In the following code snippets, x is an instance of a iSEEfgseaResults class. Refer to the documentation for each method for more details on the remaining arguments.

embedPathwaysResults(x, se, name, pathwayType, ...) embeds x in se under the identifier name. See embedPathwaysResults() for more details.

pathwaysList(x) returns the named list of pathway identifiers and associated

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval      padj   log2err         ES       NES size
#> 1: GO:0050727 0.0001721196 0.4396063 0.5188481 -0.2692978 -1.451968  414
#> 2: GO:0007059 0.0002489331 0.4396063 0.4984931 -0.2679523 -1.447983  435
#> 3: GO:0009154 0.0002693759 0.4396063 0.4984931 -0.5131816 -1.959451   43
#> 4: GO:0018108 0.0003369922 0.4396063 0.4984931 -0.2738705 -1.454841  370
#> 5: GO:0006195 0.0005147390 0.5100867 0.4772708 -0.4726531 -1.899157   52
#> 6: GO:0001895 0.0006181029 0.5100867 0.4772708  0.4108623  1.782682   80
#>                                     leadingEdge
#> 1: DNASE1,IL15,IL22RA2,MAS1,SERPINF1,IL1RL1,...
#> 2:     CCNB2,CHMP4B,CHMP6,PSMG2,TOP3A,CHMP5,...
#> 3:     NUDT4,NT5C2,PDE4B,FITM2,PDE9A,NUDT18,...
#> 4:       DLG4,IL15,CHMP6,IL22RA2,MERTK,IL11,...
#> 5:      NUDT4,NT5C2,SARM1,PDE4B,FITM2,PDE9A,...
#> 6:            TF,RP1,CLCN3,PIGR,POTEI,HSPB1,...

##
# Simulate a SummarizedExperiment object
##

se <- SummarizedExperiment()

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea",
  pathwayType = "GO", pathwaysList = pathways, featuresStats = gene_stats)
se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Access ----
##

pathwaysResultsNames(se)
#> [1] "fgsea"
pathwaysResults(se)
#> $fgsea
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.405117  0.947156 0.0943564  0.285383  1.015696        30
#> GO:0000018  0.454167  0.956463 0.0865400  0.213542  1.004666       134
#> GO:0000027  0.465812  0.960441 0.0865400  0.292569  0.987997        25
#> GO:0000028  0.396518  0.946846 0.0899861 -0.333017 -1.045081        19
#> GO:0000038  0.339658  0.937077 0.0978773 -0.289915 -1.062816        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.895277  1.000000 0.0519512  0.171200  0.823098       154
#> GO:2001258  0.738661  1.000000 0.0633597  0.209758  0.863852        59
#> GO:2001259  0.391489  0.941727 0.0962406  0.247905  1.032773        66
#> GO:2001267  0.723361  1.000000 0.0618406  0.274250  0.823930        17
#> GO:2001279  0.525050  0.969127 0.0766747  0.323991  0.942872        15
#>                          leadingEdge
#>                               <list>
#> GO:0000002      RRM2B,MEF2A,TWNK,...
#> GO:0000018      UBE2B,FANCB,H1-7,...
#> GO:0000027       TRAF7,RPL6,NLE1,...
#> GO:0000028    RPS27L,RPS28,RPS15,...
#> GO:0000038    ELOVL5,HACD4,ACOT1,...
#> ...                              ...
#> GO:2001257   KCNIP2,CASQ2,MIR212,...
#> GO:2001258 CASQ2,MIR212,MIR103A1,...
#> GO:2001259     KCNIP2,NPPA,STIM2,...
#> GO:2001267     PARK7,PIH1D1,NLE1,...
#> GO:2001279   MIR132,MIR766,FABP5,...
#> 
pathwaysResults(se, "fgsea")
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.405117  0.947156 0.0943564  0.285383  1.015696        30
#> GO:0000018  0.454167  0.956463 0.0865400  0.213542  1.004666       134
#> GO:0000027  0.465812  0.960441 0.0865400  0.292569  0.987997        25
#> GO:0000028  0.396518  0.946846 0.0899861 -0.333017 -1.045081        19
#> GO:0000038  0.339658  0.937077 0.0978773 -0.289915 -1.062816        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.895277  1.000000 0.0519512  0.171200  0.823098       154
#> GO:2001258  0.738661  1.000000 0.0633597  0.209758  0.863852        59
#> GO:2001259  0.391489  0.941727 0.0962406  0.247905  1.032773        66
#> GO:2001267  0.723361  1.000000 0.0618406  0.274250  0.823930        17
#> GO:2001279  0.525050  0.969127 0.0766747  0.323991  0.942872        15
#>                          leadingEdge
#>                               <list>
#> GO:0000002      RRM2B,MEF2A,TWNK,...
#> GO:0000018      UBE2B,FANCB,H1-7,...
#> GO:0000027       TRAF7,RPL6,NLE1,...
#> GO:0000028    RPS27L,RPS28,RPS15,...
#> GO:0000038    ELOVL5,HACD4,ACOT1,...
#> ...                              ...
#> GO:2001257   KCNIP2,CASQ2,MIR212,...
#> GO:2001258 CASQ2,MIR212,MIR103A1,...
#> GO:2001259     KCNIP2,NPPA,STIM2,...
#> GO:2001267     PARK7,PIH1D1,NLE1,...
#> GO:2001279   MIR132,MIR766,FABP5,...

pathwayType(pathwaysResults(se, "fgsea"))
#> [1] "GO"
head(lengths(pathwaysList(pathwaysResults(se, "fgsea"))))
#> GO:0000002 GO:0000003 GO:0000012 GO:0000017 GO:0000018 GO:0000019 
#>         30       1506         12          2        134          7 
head(featuresStats(pathwaysResults(se, "fgsea")))
#>      PARP1    SLC25A4       DNA2       TYMP      ENDOG       LIG3 
#>  1.0443339 -1.1403361 -1.7766884  0.8968042  0.7699708  0.7243261