The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEfgseaResults(data, row.names = rownames(data)) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

Supported methods

embedPathwaysResults(x, se, name, pathwayType, class = "fgsea", ...) embeds x in the column name of metadata(se)[["iSEEpathways"]].

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
#> 
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval      padj   log2err         ES       NES size
#> 1: GO:1904994 0.0004250137 0.8976839 0.4984931 -0.5160467 -1.933192   40
#> 2: GO:0009112 0.0006474642 0.8976839 0.4772708 -0.5337062 -1.949303   34
#> 3: GO:2000242 0.0009760581 0.8976839 0.4772708  0.4078716  1.740462   75
#> 4: GO:0043507 0.0013852782 0.8976839 0.4550599 -0.4696888 -1.787353   43
#> 5: GO:0030517 0.0014936693 0.8976839 0.4550599  0.4680446  1.822152   46
#> 6: GO:0045820 0.0016424481 0.8976839 0.4550599  0.6592612  1.922224   15
#>                                 leadingEdge
#> 1:      MIR31,ETS1,MIR21,RHOA,MDK,ALOX5,...
#> 2:     CPS1,PPAT,GART,ALDH6A1,TYMS,TYMP,...
#> 3:      WNT5A,BCL2L1,RGN,GJA1,ZP3,FBXO5,...
#> 4: FZD8,FZD10,DUSP19,FGD2,SYK,TNFRSF11A,...
#> 5:    WNT5A,KIAA0319,NRP1,MT3,IFRD1,RYK,...
#> 6:  PPARA,SLC4A1,TIGAR,STAT3,NUPR1,FBP1,...

##
# Simulate a SummarizedExperiment object
##

se <- SummarizedExperiment()

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea", pathwayType = "GO")
se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Access ----
##

# TODO: create getter function
metadata(se)[["iSEEpathways"]][["fgsea"]]
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.435993  0.939310 0.0824344 -0.285592 -1.018404        30
#> GO:0000018  0.972163  0.993293 0.0500923  0.156187  0.736101       134
#> GO:0000027  0.200431  0.916145 0.1429011  0.351018  1.188576        25
#> GO:0000028  0.448077  0.943529 0.0828962 -0.321736 -1.003994        19
#> GO:0000038  0.280669  0.916145 0.1083943 -0.304366 -1.114408        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.786792  0.969497 0.0540088 -0.184889 -0.877509       154
#> GO:2001258  0.361314  0.922195 0.0919686 -0.258168 -1.057036        59
#> GO:2001259  0.282288  0.916145 0.1075544 -0.265363 -1.099013        66
#> GO:2001267  0.504149  0.950320 0.0806388  0.324401  0.974331        17
#> GO:2001279  0.578947  0.958147 0.0721798  0.313612  0.914406        15
#>                             leadingEdge
#>                                  <list>
#> GO:0000002       MEF2A,DNAJA3,RRM2B,...
#> GO:0000018      USP51,TIMELESS,MSH2,...
#> GO:0000027         DDX28,MRTO4,RPF2,...
#> GO:0000028 LOC102724159,ABT1,RPS27L,...
#> GO:0000038     HACD2,ELOVL5,SLC27A2,...
#> ...                                 ...
#> GO:2001257        MIR1-1,SRI,PPP3CB,...
#> GO:2001258        MIR1-1,SRI,PPP3CB,...
#> GO:2001259     MIR1-1,PPP3CB,NOS1AP,...
#> GO:2001267            HTRA2,LGALS9,MMP9
#> GO:2001279      ABCD2,MIR766,MIR132,...