The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEfgseaResults(data, row.names = rownames(data)) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

Supported methods

embedPathwaysResults(x, se, name, pathwayType, class = "fgsea", ...) embeds x in the column name of metadata(se)[["iSEEpathways"]].

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval      padj   log2err        ES      NES size
#> 1: GO:0002082 0.0007304038 0.9229779 0.4772708 0.6026777 2.017953   24
#> 2: GO:1903579 0.0014973827 0.9229779 0.4550599 0.5858685 1.914885   21
#> 3: GO:0036473 0.0015119543 0.9229779 0.4550599 0.3675531 1.663104   99
#> 4: GO:0031023 0.0015342792 0.9229779 0.4550599 0.3236792 1.557857  152
#> 5: GO:0034472 0.0015591238 0.9229779 0.4550599 0.5850159 1.947904   23
#> 6: GO:0007098 0.0020956627 0.9229779 0.4317077 0.3287556 1.555042  136
#>                                      leadingEdge
#> 1: NUPR1,RHOA,COX7A2L,SLC25A23,TMEM135,PDE12,...
#> 2:      NUPR1,SLC4A1,TIGAR,ACTN3,STAT3,HDAC4,...
#> 3:    SLC7A11,FBXW7,FOXP1,OXR1,ENDOG,SIGMAR1,...
#> 4:    CENATAC,SDCCAG8,FES,CEP44,CEP152,KAT2B,...
#> 5:   INTS6L,EXOSC5,INTS1,EXOSC4,EXOSC6,INTS9,...
#> 6:    CENATAC,SDCCAG8,FES,CEP44,CEP152,KAT2B,...

##
# Simulate a SummarizedExperiment object
##

se <- SummarizedExperiment()

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea", pathwayType = "GO")
se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Access ----
##

# TODO: create getter function
metadata(se)[["iSEEpathways"]][["fgsea"]]
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.850095  0.986208 0.0507028 -0.212424 -0.765144        30
#> GO:0000018  0.932563  0.993620 0.0471642 -0.166585 -0.778645       134
#> GO:0000027  0.323970  0.948462 0.0999277 -0.319788 -1.095007        25
#> GO:0000028  0.631902  0.972686 0.0683111  0.280145  0.882289        19
#> GO:0000038  0.211832  0.924326 0.1294429 -0.316412 -1.167350        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.858300  0.986743 0.0532386  0.175064  0.846331       154
#> GO:2001258  0.731343  0.981015 0.0632191  0.209230  0.866139        59
#> GO:2001259  0.347249  0.948462 0.0965630 -0.254639 -1.062947        66
#> GO:2001267  0.541016  0.964981 0.0738053 -0.314107 -0.946295        17
#> GO:2001279  0.873747  0.987161 0.0519512  0.236442  0.690012        15
#>                        leadingEdge
#>                             <list>
#> GO:0000002 PPARGC1A,TP53,MPV17,...
#> GO:0000018   IL4,FIGNL1,WRAP53,...
#> GO:0000027  BRIX1,RPL10,MRPL20,...
#> GO:0000028    RPL38,RPS15,PWP2,...
#> GO:0000038   ACOX2,HACD4,ACSL6,...
#> ...                            ...
#> GO:2001257     KCNA1,JPH2,CDK5,...
#> GO:2001258   MIR200C,CASQ2,EPO,...
#> GO:2001259    GALR2,KCNK3,HAP1,...
#> GO:2001267      PLAUR,ST18,BAX,...
#> GO:2001279   ANXA1,CD74,MIR132,...