The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEfgseaResults(data, row.names = rownames(data)) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

Supported methods

embedResults(x, se, name, class = "fgsea", ...) embeds x in the column name of metadata(se)[["iSEEpathways"]].

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
#> 
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval      padj   log2err         ES       NES size
#> 1: GO:0018200 6.714806e-05 0.3513858 0.5384341  0.6155916  2.148837   29
#> 2: GO:0042133 4.895561e-04 0.8159664 0.4772708  0.5785129  1.976012   27
#> 3: GO:0032616 7.533103e-04 0.8159664 0.4772708 -0.6204545 -1.938916   21
#> 4: GO:0032656 7.533103e-04 0.8159664 0.4772708 -0.6204545 -1.938916   21
#> 5: GO:1902751 1.116622e-03 0.8159664 0.4550599 -0.5250240 -1.824402   31
#> 6: GO:0071375 1.215197e-03 0.8159664 0.4550599  0.2700199  1.437637  306
#>                                    leadingEdge
#> 1:   TTLL4,FOLH1B,TTLL5,AGBL4,TTLL11,NAA20,...
#> 2:      GAD1,BCHE,DBH,ALDH5A1,SLC44A4,COLQ,...
#> 3: TNFRSF21,SPHK2,LGALS9,LEF1,TNFSF4,PRKCZ,...
#> 4: TNFRSF21,SPHK2,LGALS9,LEF1,TNFSF4,PRKCZ,...
#> 5:  RRM1,CDC25A,PHOX2B,RAB11A,RRM2B,CDC25C,...
#> 6:        ADCY6,WNT1,HDAC9,GH2,FFAR3,PDE3B,...

##
# iSEEfgseaResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment()

# Package the results in a iSEEfgseaResults object
iseefgsea_table <- iSEEfgseaResults(as.data.frame(fgseaRes))

se <- embedPathwaysResults(iseefgsea_table, se, name = "fgsea")

se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Methods ----
##

## TODO