The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Slot overview

This class inherits all its slots directly from its parent class iSEEpathwaysResults.

Constructor

iSEEfgseaResults(data, pathwayType, pathwaysList = NULL, featuresStats = NULL) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

pathwayType

A character scalar specifying the type of pathway (e.g., "GO"). See embedPathwaysResults.

pathwaysList

A named list of pathways and associated feature identifiers.

featuresStats

Feature-level statistics used in the pathway analysis.

Supported methods

In the following code snippets, x is an instance of a iSEEfgseaResults class. Refer to the documentation for each method for more details on the remaining arguments.

embedPathwaysResults(x, se, name, pathwayType, ...) embeds x in se under the identifier name. See embedPathwaysResults() for more details.

pathwaysList(x) returns the named list of pathway identifiers and associated

Author

Kevin Rue-Albrecht

Examples

library("fgsea")

##
# Simulate example data
##

simulated_data <- simulateExampleData(n_pathways = 5, n_features = 100, pathway_sizes = 15:100)

pathways_list <- simulated_data$pathwaysList
features_stats <- simulated_data$featuresStat
se <- simulated_data$summarizedexperiment

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways_list,
                  stats    = features_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>      pathway      pval      padj    log2err         ES        NES size
#> 1: pathway_1 0.1207430 0.6037152 0.22798720 -0.3490347 -1.2665429   82
#> 2: pathway_3 0.6763926 0.8778409 0.07808923 -0.2900805 -0.8716798   15
#> 3: pathway_2 0.6802920 0.8778409 0.04889708  0.2488106  0.8686300   53
#> 4: pathway_4 0.8075710 0.8778409 0.07788401 -0.2080141 -0.8073750   48
#> 5: pathway_5 0.8778409 0.8778409 0.03592087  0.2070929  0.7283728   57
#>                                                              leadingEdge
#> 1: feature_14,feature_24,feature_67,feature_84,feature_28,feature_35,...
#> 2:                           feature_35,feature_75,feature_65,feature_46
#> 3: feature_61,feature_70,feature_56,feature_55,feature_92,feature_93,...
#> 4:                feature_14,feature_24,feature_67,feature_97,feature_75
#> 5:  feature_61,feature_56,feature_4,feature_31,feature_83,feature_93,...

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea",
  pathwayType = "simulated", pathwaysList = pathways_list, featuresStats = features_stats)
se
#> class: SummarizedExperiment 
#> dim: 100 8 
#> metadata(1): iSEEpathways
#> assays(1): counts
#> rownames(100): feature_1 feature_2 ... feature_99 feature_100
#> rowData names(0):
#> colnames(8): sample_1 sample_2 ... sample_7 sample_8
#> colData names(0):

##
# Access ----
##

pathwaysResultsNames(se)
#> [1] "fgsea"
pathwaysResults(se)
#> $fgsea
#> iSEEfgseaResults with 5 rows and 7 columns
#>                pval      padj   log2err        ES       NES      size
#>           <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> pathway_1  0.120743  0.603715 0.2279872 -0.349035 -1.266543        82
#> pathway_2  0.680292  0.877841 0.0488971  0.248811  0.868630        53
#> pathway_3  0.676393  0.877841 0.0780892 -0.290080 -0.871680        15
#> pathway_4  0.807571  0.877841 0.0778840 -0.208014 -0.807375        48
#> pathway_5  0.877841  0.877841 0.0359209  0.207093  0.728373        57
#>                                    leadingEdge
#>                                         <list>
#> pathway_1 feature_14,feature_24,feature_67,...
#> pathway_2 feature_61,feature_70,feature_56,...
#> pathway_3 feature_35,feature_75,feature_65,...
#> pathway_4 feature_14,feature_24,feature_67,...
#> pathway_5  feature_61,feature_56,feature_4,...
#> 
pathwaysResults(se, "fgsea")
#> iSEEfgseaResults with 5 rows and 7 columns
#>                pval      padj   log2err        ES       NES      size
#>           <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> pathway_1  0.120743  0.603715 0.2279872 -0.349035 -1.266543        82
#> pathway_2  0.680292  0.877841 0.0488971  0.248811  0.868630        53
#> pathway_3  0.676393  0.877841 0.0780892 -0.290080 -0.871680        15
#> pathway_4  0.807571  0.877841 0.0778840 -0.208014 -0.807375        48
#> pathway_5  0.877841  0.877841 0.0359209  0.207093  0.728373        57
#>                                    leadingEdge
#>                                         <list>
#> pathway_1 feature_14,feature_24,feature_67,...
#> pathway_2 feature_61,feature_70,feature_56,...
#> pathway_3 feature_35,feature_75,feature_65,...
#> pathway_4 feature_14,feature_24,feature_67,...
#> pathway_5  feature_61,feature_56,feature_4,...

pathwayType(pathwaysResults(se, "fgsea"))
#> [1] "simulated"
head(lengths(pathwaysList(pathwaysResults(se, "fgsea"))))
#> pathway_1 pathway_2 pathway_3 pathway_4 pathway_5 
#>        82        53        15        48        57 
head(featuresStats(pathwaysResults(se, "fgsea")))
#>  feature_1  feature_2  feature_3  feature_4  feature_5  feature_6 
#> -0.6264538  0.1836433 -0.8356286  1.5952808  0.3295078 -0.8204684