The iSEEfgseaResults class

The iSEEfgseaResults class is used to provide an common interface to pathway analysis results produced by the fgsea package. It provides methods to access the set of features in each pathway.

Slot overview

This class inherits all its slots directly from its parent class iSEEpathwaysResults.

Constructor

iSEEfgseaResults(data, pathwayType, pathwaysList = NULL, featuresStats = NULL) creates an instance of a iSEEfgseaResults class, with:

data

A data.frame produced by fgsea::fgsea().

pathwayType

A character scalar specifying the type of pathway (e.g., "GO"). See embedPathwaysResults.

pathwaysList

A named list of pathways and associated feature identifiers.

featuresStats

Feature-level statistics used in the pathway analysis.

Supported methods

In the following code snippets, x is an instance of a iSEEfgseaResults class. Refer to the documentation for each method for more details on the remaining arguments.

embedPathwaysResults(x, se, name, pathwayType, ...) embeds x in se under the identifier name. See embedPathwaysResults() for more details.

pathwaysList(x) returns the named list of pathway identifiers and associated

Author

Kevin Rue-Albrecht

Examples

library("org.Hs.eg.db")
#> Loading required package: AnnotationDbi
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval     padj   log2err         ES       NES size
#> 1: GO:0072583 0.0001676150 0.871716 0.5188481  0.5045286  1.977645   46
#> 2: GO:0035307 0.0003622146 0.871716 0.4984931  0.4864190  1.906659   46
#> 3: GO:2000369 0.0010836755 0.871716 0.4550599  0.6234918  1.932297   18
#> 4: GO:0060122 0.0011460103 0.871716 0.4550599  0.5312674  1.935137   33
#> 5: GO:1902369 0.0012538381 0.871716 0.4550599  0.3718872  1.631604   85
#> 6: GO:0051954 0.0016461891 0.871716 0.4550599 -0.4759904 -1.823632   43
#>                                      leadingEdge
#> 1:        AP2A2,DAB2,HIP1R,ITSN2,USH1G,SCYL2,...
#> 2:   PPP1R15B,PPP1R15A,SYMPK,PTPRC,PPIA,VRK3,...
#> 3:          DAB2,HIP1R,USH1G,SCYL2,DGKD,TNK2,...
#> 4:       CLRN2,ADGRV1,SCRIB,CDH23,CECR2,WHRN,...
#> 5: METTL16,APOBEC1,RBM46,SYNCRIP,YBX3,ELAVL4,...
#> 6:       TACR2,SLC12A2,KMO,GDNF,ADORA2A,ABAT,...

##
# Simulate a SummarizedExperiment object
##

se <- SummarizedExperiment()

##
# iSEEfgseaResults ----
##

# Embed the FGSEA results in the SummarizedExperiment object
se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea",
  pathwayType = "GO", pathwaysList = pathways, featuresStats = gene_stats)
se
#> class: SummarizedExperiment 
#> dim: 0 0 
#> metadata(1): iSEEpathways
#> assays(0):
#> rownames: NULL
#> rowData names(0):
#> colnames: NULL
#> colData names(0):

##
# Access ----
##

pathwaysResultsNames(se)
#> [1] "fgsea"
pathwaysResults(se)
#> $fgsea
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.472897  0.972715 0.0785029 -0.279001 -1.001422        30
#> GO:0000018  0.508547  0.978732 0.0817516  0.208684  0.989219       134
#> GO:0000027  0.230483  0.946059 0.1215433 -0.343579 -1.170825        25
#> GO:0000028  0.177686  0.932495 0.1492075  0.391760  1.238733        19
#> GO:0000038  0.841418  1.000000 0.0503964 -0.212254 -0.781997        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.250474  0.955391 0.1172497 -0.228621 -1.093868       154
#> GO:2001258  0.567518  0.980902 0.0679923 -0.228728 -0.944261        59
#> GO:2001259  0.788104  1.000000 0.0532386 -0.203442 -0.848034        66
#> GO:2001267  0.963532  1.000000 0.0454968 -0.200873 -0.605458        17
#> GO:2001279  0.265226  0.959147 0.1157344 -0.400062 -1.156893        15
#>                        leadingEdge
#>                             <list>
#> GO:0000002   MEF2A,DNA2,METTL4,...
#> GO:0000018    RPA2,UBQLN4,APLF,...
#> GO:0000027 FASTKD2,TRAF7,MRTO4,...
#> GO:0000028  RPS28,RRP7BP,RPS14,...
#> GO:0000038 HACD2,ABCD2,SLC27A2,...
#> ...                            ...
#> GO:2001257   KCNE1,STIM2,FGF13,...
#> GO:2001258     KCNE1,KCNE4,SRI,...
#> GO:2001259   KCNE1,STIM2,KCNC2,...
#> GO:2001267       IL9,NLE1,MMP9,...
#> GO:2001279    CD74,PTGS2,ABCD2,...
#> 
pathwaysResults(se, "fgsea")
#> iSEEfgseaResults with 5218 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.472897  0.972715 0.0785029 -0.279001 -1.001422        30
#> GO:0000018  0.508547  0.978732 0.0817516  0.208684  0.989219       134
#> GO:0000027  0.230483  0.946059 0.1215433 -0.343579 -1.170825        25
#> GO:0000028  0.177686  0.932495 0.1492075  0.391760  1.238733        19
#> GO:0000038  0.841418  1.000000 0.0503964 -0.212254 -0.781997        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001257  0.250474  0.955391 0.1172497 -0.228621 -1.093868       154
#> GO:2001258  0.567518  0.980902 0.0679923 -0.228728 -0.944261        59
#> GO:2001259  0.788104  1.000000 0.0532386 -0.203442 -0.848034        66
#> GO:2001267  0.963532  1.000000 0.0454968 -0.200873 -0.605458        17
#> GO:2001279  0.265226  0.959147 0.1157344 -0.400062 -1.156893        15
#>                        leadingEdge
#>                             <list>
#> GO:0000002   MEF2A,DNA2,METTL4,...
#> GO:0000018    RPA2,UBQLN4,APLF,...
#> GO:0000027 FASTKD2,TRAF7,MRTO4,...
#> GO:0000028  RPS28,RRP7BP,RPS14,...
#> GO:0000038 HACD2,ABCD2,SLC27A2,...
#> ...                            ...
#> GO:2001257   KCNE1,STIM2,FGF13,...
#> GO:2001258     KCNE1,KCNE4,SRI,...
#> GO:2001259   KCNE1,STIM2,KCNC2,...
#> GO:2001267       IL9,NLE1,MMP9,...
#> GO:2001279    CD74,PTGS2,ABCD2,...

pathwayType(pathwaysResults(se, "fgsea"))
#> [1] "GO"
head(lengths(pathwaysList(pathwaysResults(se, "fgsea"))))
#> GO:0000002 GO:0000003 GO:0000012 GO:0000017 GO:0000018 GO:0000019 
#>         30       1506         12          2        134          7 
head(featuresStats(pathwaysResults(se, "fgsea")))
#>       PARP1     SLC25A4        DNA2        TYMP       ENDOG        LIG3 
#>  1.10624729 -0.29232476 -1.98848862 -0.90770816 -0.79173908  0.07942219