Extract contrast results embedded in a SummarizedExperiment object

pathwaysResults returns either all pathway analysis results stored in object or a single pathway analysis result by name.

pathwaysResultsNames returns the names of pathway analysis results embedded in object.

pathwaysResultsNames(object)

pathwaysResults(object, name)

Arguments

object

A SummarizedExperiment object.

name

(Optional) Name of a single pathway analysis result name to extract. Use pathwaysResultsNames(object) to list available names.

Value

For pathwaysResultsNames: the names of embedded contrast results available.

For pathwaysResults: a list of differential expression statistics.

If name is missing, pathwaysResults returns a list in which each item contains the results of a single pathway analysis. If name is given, pathwaysResults returns a DataFrame that contains the results of a single pathway analysis.

Examples

#' library("iSEEpathways")
library("org.Hs.eg.db")
library("fgsea")
library("iSEE")

# Example data ----

## Pathways
pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
#> 'select()' returned 1:many mapping between keys and columns
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)
len_pathways <- lengths(pathways)
pathways <- pathways[len_pathways > 15 & len_pathways < 500]

## Features
set.seed(1)
# simulate a score for all genes found across all pathways
feature_stats <- rnorm(length(unique(unlist(pathways))))
names(feature_stats) <- unique(unlist(pathways))
# arbitrarily select a pathway to simulate enrichment
pathway_id <- "GO:0046324"
pathway_genes <- pathways[[pathway_id]]
# increase score of genes in the selected pathway to simulate enrichment
feature_stats[pathway_genes] <- feature_stats[pathway_genes] + 1

# fgsea ----

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
  stats    = feature_stats,
  minSize  = 15,
  maxSize  = 500)
head(fgseaRes[order(pval), ])
#>       pathway         pval         padj   log2err        ES      NES size
#> 1: GO:0046323 1.174458e-10 4.609944e-07 0.8266573 0.5982413 2.555596   75
#> 2: GO:0008645 1.840665e-10 4.609944e-07 0.8266573 0.5115897 2.367116  117
#> 3: GO:0010827 4.296846e-10 5.194149e-07 0.8140358 0.5797781 2.475614   77
#> 4: GO:1904659 4.524314e-10 5.194149e-07 0.8140358 0.5117531 2.365294  114
#> 5: GO:0046324 5.184816e-10 5.194149e-07 0.8012156 0.6314064 2.542207   59
#> 6: GO:0015749 6.491366e-10 5.419208e-07 0.8012156 0.4997745 2.325238  120
#>                                    leadingEdge
#> 1:   PTPN11,TSC1,CREBL2,PTH,MIR107,SELENOS,...
#> 2:    PTPN11,SLC2A8,TSC1,CREBL2,PTH,MIR107,...
#> 3: PTPN11,CREBL2,PTH,MIR107,SELENOS,SLC1A2,...
#> 4:    PTPN11,SLC2A8,TSC1,CREBL2,PTH,MIR107,...
#> 5: PTPN11,CREBL2,PTH,MIR107,SELENOS,SLC1A2,...
#> 6:    PTPN11,SLC2A8,TSC1,CREBL2,PTH,MIR107,...

# iSEE ---

ngenes <- length(feature_stats)
cnts <- matrix(rnbinom(n=ngenes*2, mu=100, size=1/0.5), nrow=ngenes)
rownames(cnts) <- names(feature_stats)
se <- SummarizedExperiment(assay = list(counts = cnts))

se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea", pathwayType = "GO",
  pathwaysList = pathways, featuresStats = feature_stats)

##
# List result names ---
##

pathwaysResultsNames(se)
#> [1] "fgsea"

##
# Extract results ---
##

pathwaysResults(se)
#> $fgsea
#> iSEEfgseaResults with 5009 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.720238  0.947642 0.0605093  0.238222  0.832189        30
#> GO:0000018  0.915020  0.979881 0.0491928  0.166681  0.791222       134
#> GO:0000027  0.114458  0.752723 0.1864326  0.400757  1.335463        25
#> GO:0000028  0.432540  0.888760 0.0865400  0.324662  1.006760        19
#> GO:0000038  0.280318  0.873583 0.1128434 -0.310992 -1.122412        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001256  0.746450  0.952100 0.0598603  0.273882  0.800013        16
#> GO:2001257  0.399209  0.887816 0.0908241  0.211786  1.026326       154
#> GO:2001258  0.514056  0.898606 0.0778840  0.240303  0.967522        59
#> GO:2001259  0.188377  0.817186 0.1420566  0.281733  1.172071        66
#> GO:2001267  0.458586  0.890665 0.0843144  0.330746  0.990257        17
#>                              leadingEdge
#>                                   <list>
#> GO:0000002           TYMP,RRM1,MEF2A,...
#> GO:0000018          PTPRC,TGFB1,CGAS,...
#> GO:0000027          RPLP0,RPL6,NOP53,...
#> GO:0000028 LOC102724159,RPS27L,MRPS7,...
#> GO:0000038       ELOVL1,TECR,SLC27A6,...
#> ...                                  ...
#> GO:2001256         SARAF,SLC8B1,JPH4,...
#> GO:2001257          ANK2,KCNC2,GSTO1,...
#> GO:2001258      GSTO1,PPP3CC,MIR26A1,...
#> GO:2001259          ANK2,KCNC2,GSTO1,...
#> GO:2001267            BAX,NLE1,TFAP4,...
#> 
pathwaysResults(se, "fgsea")
#> iSEEfgseaResults with 5009 rows and 7 columns
#>                 pval      padj   log2err        ES       NES      size
#>            <numeric> <numeric> <numeric> <numeric> <numeric> <integer>
#> GO:0000002  0.720238  0.947642 0.0605093  0.238222  0.832189        30
#> GO:0000018  0.915020  0.979881 0.0491928  0.166681  0.791222       134
#> GO:0000027  0.114458  0.752723 0.1864326  0.400757  1.335463        25
#> GO:0000028  0.432540  0.888760 0.0865400  0.324662  1.006760        19
#> GO:0000038  0.280318  0.873583 0.1128434 -0.310992 -1.122412        35
#> ...              ...       ...       ...       ...       ...       ...
#> GO:2001256  0.746450  0.952100 0.0598603  0.273882  0.800013        16
#> GO:2001257  0.399209  0.887816 0.0908241  0.211786  1.026326       154
#> GO:2001258  0.514056  0.898606 0.0778840  0.240303  0.967522        59
#> GO:2001259  0.188377  0.817186 0.1420566  0.281733  1.172071        66
#> GO:2001267  0.458586  0.890665 0.0843144  0.330746  0.990257        17
#>                              leadingEdge
#>                                   <list>
#> GO:0000002           TYMP,RRM1,MEF2A,...
#> GO:0000018          PTPRC,TGFB1,CGAS,...
#> GO:0000027          RPLP0,RPL6,NOP53,...
#> GO:0000028 LOC102724159,RPS27L,MRPS7,...
#> GO:0000038       ELOVL1,TECR,SLC27A6,...
#> ...                                  ...
#> GO:2001256         SARAF,SLC8B1,JPH4,...
#> GO:2001257          ANK2,KCNC2,GSTO1,...
#> GO:2001258      GSTO1,PPP3CC,MIR26A1,...
#> GO:2001259          ANK2,KCNC2,GSTO1,...
#> GO:2001267            BAX,NLE1,TFAP4,...